backend/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "petgestor-backend"
version = "0.1.0"
description = "FastAPI backend for PetGestor Digital Hub"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy[asyncio]>=2.0.30",
  "asyncpg>=0.29.0",
  "aiosqlite>=0.19.0",
  "alembic>=1.13.1",
  "pydantic>=2.7.1",
  "pydantic-settings>=2.2.1",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "email-validator>=2.1.0",
  "pyjwt[crypto]>=2.8.0",
  "python-multipart>=0.0.9",
  "stripe>=10.0.0",
  "fastapi-mail>=1.4.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.6",
  "httpx>=0.27.0",
  "ruff>=0.4.5"
]

[tool.ruff]
line-length = 99
respect-github-ignores = true

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

